requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
neo4j==5.24.0
meilisearch-python-async==1.8.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
httpx==0.27.2